refactor(db) 11/13: db/passes.rs explicit pipeline + slim main.rs (closes #257)
Move the ~45-call derived-table populate sequence (verbatim, with per-pass
prints and the load-bearing ordering comments) out of main.rs into
db::passes::run_passes(&Database, &PassCtx) -> PassCounts. PassCtx carries the
filesystem inputs the world passes need; PassCounts returns the 18 counts the
run summary reports. main.rs drops 1320 -> 989 lines. No graph engine -- a
documented linear sequence. 240 tests + clippy -D warnings + fmt green.
